+++ packages/web-components/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.61.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.60.0...@carbon-labs/web-components@0.61.0) (2025-09-17)
+
+
+### Features
+
+* **wc-global-header:** Contribute new Global Header component ([#774](https://github.com/carbon-design-system/carbon-labs/issues/774)) ([e8b4677](https://github.com/carbon-design-system/carbon-labs/commit/e8b4677780d68f8ccf02860e327fcaa1d7a3c833))
+
+
+
+
+
+# [0.60.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.59.0...@carbon-labs/web-components@0.60.0) (2025-09-16)
+
+
+### Bug Fixes
+
+* stylepicker replace icons ([#806](https://github.com/carbon-design-system/carbon-labs/issues/806)) ([b8583ad](https://github.com/carbon-design-system/carbon-labs/commit/b8583ad57d6721642a3a9b9a99782953cc62cd3b))
+
+
+
+
+
 # [0.59.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.58.0...@carbon-labs/web-components@0.59.0) (2025-09-16)
 
 **Note:** Version bump only for package @carbon-labs/web-components